Make snapping over small holes work with endpoints
When the point being snapped to is an endpoint (or any coordinate that
defines a line) rather than interior point to a line, it is better to
use exactly that point. If however it is interior to the line, it is
better to move the point just across the line to ensure it really does
cross when reprojected etc.